+++ elmyra.ip.access.epo/CHANGES.rst
@@ -9,6 +9,7 @@ development
 - ui: fix search interface behavior weirdness
 - ui: medium refactoring of main/core javascript code to components/*
 - ui: refactor ops-fulltext specific code from core.js to ops.js
+- middleware/ui: retrieve german fulltexts from DEPATISconnect
 
 0.37.0
 ======